requirements.txt
haystack
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
pypdf
sentence-transformers
-e .